starjava: enable global build in absence of JAI
Modify the top-level build.xml file so that if the JAI extension
classes are absent at build time, the JAI-dependent packages are
simply ignored, and the non-JAI-dependent packages are built
successfully.  Previously, in the absence of JAI the whole build
terminated midway (after attempting to build JSKY) with a fatal error.

Where JAI is present at build time, behaviour following this commit
should be more or less unchanged.  There are some minor differences:
packages are built in a different order, and JAIUTIL and JSKY
javadocs are no longer built by the javadocs-full target.

To restore (approximately) the previous behaviour, which is for
the global build to report a fatal error in absence of JAI,
you can set the jai.present property, e.g. "ant -Djai.present=true build"
(though note that -Djai.present=false doesn't do the opposite -
setting that property to any value has the same effect).
You might want to do that if you're building JAI-dependent packages
(e.g. SoG, SPLAT), so that a failure is more obvious when JAI is missing.

The packages that will, following this commit, not be built
in absence of JAI are JSky, JAIUTIL, SoG, SPLAT, Treeview and FROG.
JAI is essential at build and run time for JSky and hence SoG;
it is required at build time for SPLAT, but I don't know what
parts of SPLAT actually use it or how easy it would be to remove
that dependency.  FROG is dependent on parts of SPLAT, though
probably not the JAI-dependent parts (though as far as I know,
nobody uses FROG).  Only small parts of Treeview are JAI-dependent,
they could probably be extricated.

Java Advanced Imaging (JAI) is a set of interfaces with an associated
reference implementation provided by Sun Microsystems.
Last updated in 2006, its functionality has been superseded by ImageIO.
It is licensed under the "JDL", the JAI (or Java?) Distribution Licence,
which by my reading does not make it suitable for distribution
as part of starjava software.  It is at time of writing still
available from Oracle at
https://www.oracle.com/java/technologies/advanced-imaging-api.html

Up till now there has been a requirement that anybody building
starjava has JAI installed in their JDK.  But it's not very
straightforward to do that, meaning that people were in general
unable to build even non-JAI-dependent parts of starjava.
This commit enables a build of the non-JAI-dependent parts
(e.g. TOPCAT) when JAI is absent.