ttools: avoid SunDoclet for java >=9
For JDK versions > 1.8, do not attempt to build or use the class
uk.ac.starlink.ttools.build.SunDoclet
(which is deprecated at Java 9 and removed at Java 17), use
uk.ac.starlink.ttools.build.JdkDoclet instead.

The hard work here was done a year or so ago when JdkDoclet was
written as a SunDoclet drop-in replacement; that enabled builds
with JDKs <17.  This commit completes the job by putting logic
in the TTOOLS and TOPCAT build files to make sure that JdkDoclet
is actually used in preference to SunDoclet where possible,
and means that the build can be done for JDK 17 and later.