ant: hack ant to avoid SecurityManager usage
JDK17 deprecates use of java.lang.SecurityManager, and at JDK21
System.setSecurityManager() throws an UnsupportedOperationException.
Under certain circumstances, Ant uses System.setSecurityManager:
the <java> task manipulates the current JVM's SecurityManager so
that Permissions different from the parent process can be set.
This happens unless the fork attribute is set true.

We could work round this by settting fork="true" on all <java>
task invocations in the build files, but that would be somewhat
inefficient and might(?) have other knock-on effects.
Instead, make a small change to the Ant <java> task implementation
so that setSecurityManager is never invoked, unless the <java>
task element has an explicit <permissions> child (which is not the
case anywhere in starjava).

It looks to me like the effect of this change will be nil; but in
any case I'm pretty sure it's harmless for our purposes since we
don't attempt any security cleverness in the starjava build.

I expect that later versions of Ant have addressed this, but I don't
want to mess with the build system by upgrading Ant if I can help it.

Note this is not a permanent fix: since SecurityManager is
@Deprecated(since="17",forRemoval=true), presumably the class
will disappear altogether in some future JDK release and
wreak more mayhem at ant build and possibly run time.
But we'll worry about that when it happens.