topcat: avoid use of SecurityManager
java.lang.SecurityManager is deprecated for removal since JDK 17,
so remove references to it.

The SecurityManager was being used here to determine whether
local disk read and write would be permitted, and hence whether
to disable relevant Load/Save dialogues.
In some unsigned WebStart contexts it used to be the case that
local disk access might be blocked by the sandbox,
but since WebStart, especially unsigned, is not really used or
supported any more, I don't think there's much point in checking
these things.  It is therefore assumed from now that local disk I/O
is, in principle, possible.