topcat: pre-build JavaFX-based internal browser
The FxHtmlPanel is a concrete subclass of AbstractHtmlPanel,
based on JavaFX, that does a better job (more capable, less ugly,
javascript-friendly) of rendering HTML than the other supplied
concrete subclass, SwingHtmlPanel.  One or other of these subclasses
is used for internal display of web pages, e.g. as an activation action,
depending on whether JavaFX is available at runtime.

JavaFX was effectively bundled only with version 8-10 of Java,
being dropped from later versions of the Java SE.
That means that (a) building the FxHtmlPanel class is hard to
support for JDKs >=11, and (b) it's not very likely that users
will find themselves with JavaFX available at runtime, so few will
benefit from this feature.

Enabling a build of the FxHtmlPanel class that works for JDKs >=11
would therefore be a bit of a pain.  So at this commit,
the FxHtmlPanel class is removed from the buildable source code.
However, a jar file containing a pre-built copy of the class
(compiled using a standard Oracle JDK 8), along with the
java source file, is provided as topcatfx.jar.
No attempt is made to build this jar at build time.

So if JavaFX is present at runtime the JavaFX-based browser
will be available, otherwise only the Swing fallback can be used,
and there is no additional build-time complication or requirements
to build in a JavaFX-capable environment.

I think that the runtime invocation is protected by enough checks
that it won't cause runtime exceptions if invoked in surprising
environments.  But if it causes trouble, this functionality
could just be removed altogether, it won't benefit that many users.

Alternatively javafx classes could in principle be included in
the starjava sources and bundled with the application;
the OpenFX SDK (GPL) is available from
https://gluonhq.com/products/javafx/, specifically
https://download2.gluonhq.com/openjfx/11/openjfx-11_linux-x64_bin-sdk.zip
From the javafx/sdk-11/lib/ dir, the jar files
javafx.base.jar javafx.graphics.jar javafx.web.jar
(about 5Mb total) would be required at compile- and runtime.