ttools: MethodBrowser copes better with missing items
Make a couple of changes to the MethodBrowser to make it behave
a bit better if the documentation content is missing at runtime
for a class member.  This shouldn't be the case, but can happen
for a small number of fields (representing case-folding collisions,
e.g. Fluxes.JOHNSON_AB_i and Fluxes.JOHNSON_AB_I) if the build
has been done on a case-insensitive filesystem.