auth: makeConnection now recurses on 3xx
There was a tacit assumption in the AuthManager code that in case
of HTTP 3xx redirection the chain of redirected URLs would all
use the same authentication context.  There's no guarantee that
this must be the case, and in fact for the current Rubin TAP
service it is not.

So instead of iterating over redirects with the same auth context,
recurse to do all the auth negotiation from scratch when a 3xx
is encountered.