ttools: accommodate VOUnit-compliant angle units
In various places make sure that the VOUnit-compliant "deg and "rad"
are accepted alongside the ad-hoc "degree(s)" and "radian(s)".

This commit does not attempt to excise the ad-hoc usages from all
the code, since it's a bit tricky to determine when these strings
will potentially end up in VOTable metadata (where VOUnit is
preferred) and when it's just user-facing.  But hopefully following
this commit, most output VOTables will not have been assigned
non-VOUnit-compliant unit attributes, and if future changes
are needed to make that happen they are less likely to break things.