topcat: disable action forwarding during table change
In the FormLayerControl, temporarily prevent the action forwarder
from receiving events while the table is changed.

An alternative way to do it would be to have a flag on ActionForwarder
that stops it forwarding events.

Table changing can cause a huge cascade of events when there are
many column selectors (this may be partly sloppy programming),
which means that changing between tables with similar columns
can take a long time because of GUI events.
Example: changing between two similar tables in the Corner plot
with 10 coordinates takes about 6 seconds of time on the EDT.

Following this change, the EDT time is not significant.
I don't *think* there should be any unwanted changes in UI behaviour
as a result.

There are (many) other places where something similar could be done,
but in most of them the number of events is not enough to cause
noticeable trouble.