starjava: add -quiet flag to javadoc invocations
The <javadoc> tasks in the package build files now have -quiet,
which suppresses information about what source files javadoc is reading.

At java 8 it just reported the packages being javadoced, which was
usually a few lines per javadoc invocation, but at Java 21 it
reports every source file, which can be hundreds of lines long and
swamps the rest of the output, which is a nuisance.
So suppress this output.

Warnings and errors are still reported.