coco: don't be an applet
Applets are obsolete, and the javax.swing.JApplet class is
deprecated for removal since Java 17.
This commit does a not-particularly-tidy job of removing any
attempt to extend JApplet, leaving the JFrame functionality,
so that attempts to build it in a JDK where Applet is actually
removed will not fail.

Probably nobody is using or will use this package anyway.