diva: don't be an applet
Applets are obsolete, and the javax.swing.Applet class is
deprecated for removal since Java 17.

This commit simply removes the AppletContext class which provides
Applet functionality, and any other classes depending on it.
Most of those depending classes are demos, which are anyway
only a thin wrapper around the corresponding application-based demo.

Future attempts to build this package in a JDK where JApplet has
actually been removed will now not fail for that reason.