ttools: remove finalizer in TapResultReader
Finalizers are evil and deprecated for removal in Java 21.
The finalizer in this case was deleting the UWS job from the
server when the table was no longer in use.
This commit just puts the relevant code into the close() method instead.
But in fact even that will rarely be triggered, since there are other
measures in place to perform this deletion (in the case that deletion
is intended).