util: add Cleaner class
This class is a drop-in replacement for java.lang.ref.Cleaner,
which is introduced at Java 9.  This class provides the same
functionality from Java 8, and can be retired if the target
platform for this package advances to Java 9.

It comes with a testcase, but the actual test in CleanerTest
is effectively commented out, since it can't be made to run
reliably because it relies on unguaranteed behaviour of the
JVM garbage collector.  But it can (maybe) be run by hand to
see if it's apparently working; this has been done, and it is.