table: don't use Float for MRT files
Clever code in MrtReader attempted to identify some column types
as 32-bit float instead of 64-bit double precision if the field
size was small enough that only float precision could fit in.
But this forgot that representations like "1e46" can use few characters,
but not fit in a float because of the large absolute value of the
exponent (<~38 for float).
So now use double precision for all floating point columns.

This problem was reported by Katie Merrell from AAS.