parquet: don't attempt to open remote files by default
Parquet files referenced by a URL (e.g. http or https) are now
by default rejected.  Previous behaviour was to attempt to open
them, which the parquet-mr library permits, but then fails at
least in the default configuration with an error message that
confuses users.  A switch is provided to turn the old behaviour
back on, using the tryUrl config option in the ParquetTableBuilder.