oldfits: avoid URL constructor deprecation warnings
URL constructors are deprecated at Java 20.

I tried to remove URL constructors from this package in favour of
URI-based techniques, but the package (which is itself more or less
deprecated) encourages use of trailing square brackets to reference HDUs
(e.g. "http://example.com/foo.fits[3]" for HDU#3).

Since these square brackets are not legal URI characters any such
attempt is fraught with difficulty and likely to cause regression issues.
So settle for localising the use of URL constructors as much
as possible, and suppress the deprecation warnings.