vo: fix bug in TAP curl logging
The CONFIG-logged curl equivalent commands for TAP queries were
not quite right: in the case of multipart forms it was reporting
use of --form arguments when it should have been --data (or similar).
This had me scratching my head for quite a while when I couldn't
reproduce a STILTS failure with (what I thought was) the
corresponding curl command.