votable: cope with (most) SI prefixes in TIMESYS units
For FIELDS/PARAMs referencing a TIMESYS element, an attempt is made
to assign a TimeDomain if the unit is recognised.
Till now, only seconds, day and year were recognised.
This commit still doesn't do proper VOUnit parsing of the unit
attribute, but it now looks for obvious SI previxes on seconds,
which may be found in practice: ms, us, ns, ps and ks.