ttools: update screenshot following MULTISHAPE changes
The change in the way that STCS-S UNION shapes are handled
a few commits back (90f7a382801e3) altered the way that the nominal
center of an Area is calculated for these inputs: it now takes
the mean position of the centers of any constituent shapes
(e.g. POLYGONs) rather than the mean position of all the vertices
in all the constituent polygons.  The new recipe is not obviously
better or worse than the old one, but they can produce slightly
different results, so the example plot that serves as a regression
test is updated here in accordance with the new behaviour.