ttools: allow adaptive use of tick label orientation
When requesting ticks on an axis, you now supply a list of acceptable
orientations rather than a single orientation.  Then if the tick
marks are too crowded with the first choice orientation, an attempt
is made to paint them with the second choice orientation etc.
The effect is that if labels are too crowded, a more compact
orientation (e.g. angled labels) can automatically be used.

This system is however flawed for general use, since it is prone to
come up with incorrect values for axis label space requirements
(PlotPlacement.calculateDataInsets gets the wrong value).
This is because the space requirement can change dramatically between
approximate and actual data bounds during the iterative space
requirement assessment process if the selected orientation changes.
So it works OK when that calculation is not very significant
(e.g. CubeSurface), but not when it is (e.g. PlaneSurface,
where the plot bounds are determined by working out the required
surround).

At this commit, supplied orientation lists only have a single value,
which is the value that was previously hard-coded.
So there is no behaviour change.