ttools: axis label orientations now configurable
Add a new config key for OrientationPolicy to all plots except Sky,
so that the angle of numeric axis labels can be controlled by the user.

By default the orientation is fixed at Horizontal for the Plane and
Time plots, Angled for the Matrix plot (where labels are likely to
overlap, including between plot cells) and Adaptive for the Cube plot
(which doesn't need to know its surround requirements to
position the axes).

Since the default value is different for Matrix and Plane plots,
the actual ConfigKey instance must be different for those two cases.
The fact that those two cases share the same PlotType class PlanePlotType
complicates the implementation a bit.