auth: fix 3xx relative Location bug
HTTP 3xx redirects can contain an absolute or relative URI in
their Location field: in recent versions only the absolute case
was being handled.  The effect was that redirects with relative
URIs were not being followed.  Fix it so that relative locations
are resolved against the base URI as well.

This appeared as a regression bug in some places since the AUTH
library was adopted for most redirect handling by VO and other
packages about a year ago, and before that the redirect handling
worked correctly.