topcat: improve ColumnConverter for use with stilts
Add a new method convertExpression to ColumnConverter,
and provide a new subclass AngleColumnConverter which knows about
angles.  These can be used to provide JEL expressions for constructing
STILTS command lines, based on the state of converter components
in the GUI.

The angular one is required, and more complicated than you might
expect, especially because angle value selectors in topcat in most
cases yield radian values, and we want degrees or similar to pass
to stilts.  This use of radians internally in topcat was an early,
bad, design decision.