ttools: deconfuse stilts command formatting
Refactor the classes that take StiltsCommand instances and format
them to text etc.  There was some confusion concerning function here,
and the whole thing was also tied to plot commands only.

The TTOOLS StiltsCommand and CommandFormatter classes can now be used
for working with non-plot commands.  At this commit, the relevant
topcat classes are still tied to plot commands, but it will now
be easier to disentangle that functionality.