util: adjust test case for locale robustness
Change the way that a string is specified in the PrimitiveListTest
test case.  The previous source code caused test failures in some
contexts; I think this is because the test compilation interpreted
the string literal in the source code in a locale-dependent way,
so in some locales the string literal ended up meaning something
different than what was intended.  This change makes it explicit
what unicode code points are in the string to test, which ensures
the test is doing what it's supposed to.

This problem was reported by Ole Streicher.