votable: add VOTableWriter resourceType attribute
This can be used to add the RESOURCE/@type attribute when producing
VOTable output.

This was originally introduced to facilitate fixing the resource
on VOParquet DATA-less TABLE elements, but we eventually decided
not to do it that way, so it's currently unused.  But it might
come in useful for somebody one day.