ttools: support Area.Type.MOC in sphere plot
By an oversight, the MOC arm was omitted from the switch statement
that does Area.Type-specific shape coordinate handling for the
sphere plot (why isn't there a way to write switch statements
so the compiler warns you about this??)
That resulted in a NullPointerException or AssertionError
if you tried to plot UNIQ or ASCII-MOC shapes in the sphere plot.
Fix it.

Plotting MOCs on a sphere is a bit of a strange thing to want to do,
but there's no reason it can't/shouldn't work.