ttools: add isometric config option to cube plot
If selected, it forces scaling to be the same on the X, Y and Z axes.
This is especially useful if you're using the cube plot to view
objects in real space, like the vertices or faces of an asteroid
3-d model.

Note the same thing in the Plane plot is currently called Aspect Lock
(though the UI is the same, the parameter available from stilts has
a wider applicability).