parquet: add comment about file format and version
Add a link in the user documentation about where the parquet file
format is defined and the version of that file format (2.10.0).

Note that the format definition link is not where you might expect
from a google search at time of writing;
https://parquet.apache.org/docs/ currently points at a version that
is a couple of years out of date and is missing important information.