votable: roll back lazy metadata construction change
Commit 1c6e494b27b modified VOStarTable to construct its column
and table metadata at construction time rather than lazily.

It turns out this was a bad idea, since the required information
is not available at construction time in at least certain
circumstances, for instance when streaming multiple tables from
multiple TABLE elements in a VOTable document (MultiTableStreamer) -
the elements feeding into the metadata may not exist until
RESOURCEs after the end of the TABLE are encountered.

This commit reverses the change at that commit, and adds an explanatory
comment to prevent me from doing the same thing again in the future.
The message from the earlier commit said:

   "The saved time was not significant and this will make it easier
    to modify that metadata as required by other structures in the
    VOTable document."

Saved time was not the point, but I didn't realise that when I
made the change.  The other part of this comment worries me -
have I modified VOStarTable metadata in a way that relies on
this change being in place?  I don't recall and can't see any
instances of that, but it's possible I've broken more things
by this regression.
I think I could change it so that the ColumnInfos are put in
place at construction time, but it looks pretty difficult to do that
for the table parameters.