topcat: fix boxing/conditional regression bug
When making changes to modernise the source code (accommodate
java 21's distaste for use of primitive wrapper type constructors
like new Integer(i)), I made some 'simplifications' of use of
conditional operators by removing what I thought were redundant
casts (required at java 1.4, but no longer required at 1.5,
since the ternary operator at 1.5 can automatically work with
second and third operands of different types by returning a value
that is the least common supertype of these two operands).

Thus, as well as the new Integer() -> Integer.valueOf() changes
at those commits, I removed the (Number) casts from expressions like:

   isInt ? (Number) Integer.valueOf((int)x) : (Number) Double.valueOf(x);

since when these were written the (Number) casts were required to
make the expressions compile, but now they are not.
But it turns out that this change was not harmless, because owing to
(I think) an unholy combination of conditional expression result type
rules and unboxing semantics, the runtime result of the following
expression:

   Object zero = true ? Integer.valueOf(0)
                      : Double.valueOf(0);

is, very surpringingly to me, a Double.  The return type of this one:

   Object zero = true ? (Number) Integer.valueOf(x)
                      : (Number) Double.valueOf(x);

is Integer.  So I've reinstated those casts.

One noticeable effect of the previous state of this code was that
the Count column of the imported histogram table in topcat was broken
(declared as Double but with Integer content) and could not be used
in most contexts.  Thanks to Alexey Mints for reporting this.