parquet: fix major error in array-valued input
The previous reader mostly ignored blank array elements in
the values of array-valued columns.  This was due to a misunderstanding
of the parquet format, which it took quite a while to work through.

At this commit however there are still issues with blank array values.
Comment out the parquet tests in the TABLE package since they're
failing; they will be brought back once these issues are ironed out.