parquet: enforce correct use of groupArray option
If a ParquetTableWriter with groupArray=false is used to write
data that contains null arrays or arrays with null elements,
an IOException will be thrown.

Since the relevant parquet-mr library methods don't declare any
checked exceptions, this gets thrown as a typed RuntimeException
and then caught higher up and wrapped as an IOException there.
That typed exception class is considered an implementation detail
so is currently package-private.