xdoc: make toHelp.xslt aware of subdiv element
When I introduced subdiv in 2022, I fixed the toHTML.xslt stylesheet
to process it and its children, but not toHelp.xslt.
The consequence was that any subdiv elements in SUN/253 showed
up OK in the HTML documentation, but not in the in-application
help browser.

Fix that here.

Spotted by Gabriel Bihain.