ttools: upgrade Moc library to v6.31
Replace Moc.jar with a version downloaded from
https://wiki.ivoa.net/internal/IVOA/MocInfo/JavaMoc2.jar
(r2 of that file), which claims to be v6.31, Dec 2024.
The same jar file is available from
https://github.com/cds-astro/cds.moc/blob/master/Moc.jar
at commit 7ab4ad5; the sources at that commit presumably
correspond to the same file.

This is supposed to be MOC2.0 compliant.

The HealpixMoc class is effectively replaced by the SMoc class.
There are some changes of API.