ttools: add home-grown MOC implementation classes
Add interfaces and implementations for building MOCs by going through
lists of HEALPix tiles.  There is an implementation based on
the CDS MOC library and another one built from scratch.

This code, in the new package uk.ac.starlink.ttools.moc,
is somewhat experimental.  In some cases I'm not so happy with the
memory and CPU usage of the CDS implementations, so I wanted to
see how far I could get doing it on my own, which is probably better
for lower maximum orders, though not good at higher ones.
The home-grown implementation may not be as well tested as the CDS one,
and may have bugs.  The idea is partly to provide a framework for
comparing performance and results of the different implementations.