topcat: generalise layer control override mechanism
Provide a hook in the LayerControlAction interface that allows an
instance to jump the queue when offering to plot a given table.
Previously, this was done with an ugly hack for only a single case,
the Healpix plotter (if there was a table that looked like HEALPix
data).  That logic is now moved inside the HealpixLayerControl,
and there is the opportunity for other layer controls to do something
similar.

There should be no behaviour change at this commit.