table: add VerTableBuilder
New input handler for .ver format tables.  These represent 3d shape
models e.g. for small solar system bodies.

This is hacky because the formats aren't well defined, but it works
for a few tables I've tried it with.

Should be extended one day to deal with better documented vertex file
formats, as described in the MKDSK docs, currently available from
https://naif.jpl.nasa.gov/pub/naif/utilities/PC_Linux_64bit/mkdsk.ug

Given the general hackiness of this, it is not advertised as a STIL
input format.  But you can use it by setting the system property
-Dstartable.readers=uk.ac.starlink.table.formats.VerTableBuilder

The testcases provided read some example ver files (they are probably
not all properly called ver), which can be parsed by the current version
of the VerTableBuilder, and are adapted from the example data above.
I don't really now what the name of each of the format variants is,
or whether one can expect to find other examples of these formats
elsewhere.

At this commit the testcase examples are:

   rosetta.ver:
      based on http://comsim.esac.esa.int/rossim/SHAPE_MODEL_DRAFTS/OTHER_LEGACY/SHAP7_v1.6/cg-spc-shap7-v1.6-cheops.ver
      Corresponds to MKDSK file type 4,
      labelled there "Rosetta/OSIRIS ``ver'' file"

   damit.ver:
      based on https://astro.troja.mff.cuni.cz/projects/damit/stored_files/open/60970/shape.txt

   gaskell.ver:
      based on https://sbnarchive.psi.edu/pds4/non_mission/gaskell.phobos.shape-model/data/phobos_ver64q.tab
      This is *not* identical to MKDSK file type 2,
      labelled there "Gaskell Shape Model"

The following generate some example plots for the 3-d bodies described
by the files at URLs referenced above:

   stilts plot2cube \
          isometric=true frame=false zoom=2 auxvisible=false \
          in=phobos_ver64q.tab#2 ifmt=ver \
          aux='hypot(x,y,z)' shading=aux auxmap=cividis auxflip=true \
          x=X y=Y z=Z otherpoints=others \
          layer1=polygon polymode1=fill

   stilts plot2cube \
          isometric=true frame=false zoom=2 auxvisible=false \
          in=cg-spc-shap7-v1.6-cheops.ver#2 ifmt=ver \
          x=X y=Y z=Z otherpoints=others aux='hypot(x,y,z)' shading=aux \
          layer_1=Polygon polymode_1=fill

   stilts plot2cube isometric=true frame=false zoom=1.5 auxvisible=false \
          in=shape.txt#2 ifmt=ver x=x y=y z=z otherpoints=others \
          aux='hypot(x,y,z)' shading=aux layer1=polygon polymode1=fill