topcat: fix DomainMapperComboBox model reselection
When the model of a ColumnDataComboBox changed, its associated
DomainMapperComboBox (if any) wasn't getting messaged at the right
time, with the effect that it ended up getting reset to null rather
than to a suitable default for the new selection.
One consequence of this was that when changing table in the
TimePlotWindow the time domain mapping selector for the time
coordinate got reset to null rather than configured to a suitable
default value.  Fixed by explicitly messaging the domain mapping
selector on column data model change.