ttools: use non-threadsafe RowComparators in SortFilter
Modify SortFilter so that it uses RowComparator objects that
are confined to a single thread, in conjunction with the
SortUtils method that will work with non-threadsafe IntComparators.
This avoids use of ThreadLocals in RandomJELEvaluator, which ought
to speed things up or at least uncomplicate them a bit.
However JELRowReader still has synchronized evaluate methods,
so this may not have much effect in practice.