ttools: accommodate GPL-encumbered classes
The build now produces a file ttools-gpl.jar alongside ttools.jar,
containing only those classes in uk.ac.starlink.ttools.gpl,
which is intended as a home for classes that are encumbered with
the GNU General Public Licence.

Inclusion of such files isn't going to make much difference to users of
the software, since STILTS/TTOOLS is already heavily dependent on JEL,
which is GPL.  But the LICENSE.txt file notes that the code in
ttools.jar is LGPL, which may be of use to people who want to
grab parts of the code and reuse it.

This new ttools-gpl.jar file is noted as GPL in the LICENSE.txt file.