parquet: include zstd-jni library
The zstd-jni library is required to read parquet files using the
ZSTD compression algorithm.  Gaia DR4 may use ZSTD, and possibly
other projects as well.  The jar file is another 6Mb, mostly made
up of native libraries (14 of them).
The file in question is available from the maven central repository at
https://repo1.maven.org/maven2/com/github/luben/zstd-jni/1.5.0-1/zstd-jni-1.5.0-1.jar

It's bad enough having to package this with the application,
I don't want to commit it to the repository as well, so download
it from the central repository during the build process,
without deleting it in the cleanup phase.  Something similar
is done for jython in the TTOOLS package.

There are already a lot of parquet-mr jar files in the repository
(see src/lib).  If I come to update these at some point, I should
probably use this download technique rather than storing them in
the repository, but there's no benefit in doing it before they
need changing.