topcat: rewrite prolematic PNG
When generating the PDF version of the user document (ant docs-pdf),
Fop objected to the ControlWindow.png image citing an "Image Error".
I've had this before with PNGs that include thumbnails, but this
one didn't seem to, so I don't know what the problem was.
But I fixed it by cut'n'pasting it into a new image and resaving
in GIMP.