ttools: write Scale variants of PlotUtil isLog functions
For those functions in PlotUtil with a boolean isLog parameter,
rewrite them in terms of a Scale parameter.
For now, leave in place implementations with the old signatures,
but based on the new implementations.

This causes some regression features in the plots generated by the
plot2-examples target.  The effect is that on logarithmic axes,
an annotated tick mark exactly at one end of the axis is now sometimes
painted when it wasn't before.  I think this is more correct behaviour,
so include the changes.