ttools: fix problem with taplint upload table
The table used for uploads in taplint's UPL stage had questionable
metadata - arraysize="2" for string values that were in some cases
one character long.  It was then requiring that round-tripped values
came back with the same actual length, rather than declared length,
as the strings sent, otherwise triggering an E-UPL-TMCD error.
This was at least unfair, and perhaps incorrect.

Fix it so that columns with variable array content are labelled
arraysize="*".