ttools: rewrite Axis to use Scale
Axis is now a concrete class with a Scale member rather than an
abstract class with different implementations for Linear and
Log scaling.

I didn't expect any change of behaviour but there are a couple of
pixels different in one of the test plots.  I think this is
rounding errors or something.