ttools: replace log by scale config in plot UIs
The various SurfaceFactories now declare Scale ConfigKeys instead
of Log ones, so that scales other than linear and log can be selected.

Actually that's not quite true: the log keys are still present in
the declared lists of profile config keys for backward compatibility
purposes, so that legacy scripts with settings like "xlog=true" will
still work.  But these keys are marked deprecated in the documentation,
and they are not visible in the topcat GUI.

It is therefore possible in stilts (but not topcat) for the user
to present non-default values for both the scale and the log options,
which would indicate some confusion on their part.  In this case a
warning message is written through the logging system.  This lack of
orthogonality is a bit suboptimal, but maintaining the backward
compatibility is important enough to justify it.

There is no change to the topcat MatrixAxesController, so you can't
use the new scalings for a corner plot in topcat.  That's because
the GUI is already very crowded and I don't for now want to attempt
replacing the existing compact log checkboxes (already not so easy
to use here) with the more flexible scaling selection UI for multiple
axes in the relevant component.  If somebody asks I might revisit this.