ttools: add asinh Scaling option for aux scales
The aux scaling selection now adds asinh to the existing options.
This is a bit unlike the other ones since the zero point of the
input data is tied to a particular point in the scale rather than
being assigned from the input data, and the normalisation of the
scale is chosen somewhat arbitrarily.  The labelling of the aux
axis is currently done linearly, which doesn't represent the
scale very well.

Some adjustment may be desirable in future, but it manages to represent
a wide dynamic range of data (like log) that may be positive or negative
(unlike log).