topcat: disallow overwrite of standard RowSubsets
Prevent users from supplying new definitions of the standard
RowSubset instances, All and Activated.

This does interfere slightly with the working of TopcatCodec1, which has
grown a hack to work round it and which now will not persist the Activation
subset index between sessions.  But since the v1 session codec is more or
less obsolete (only present for backward compatibility with old session
files) and activation status is not very important between sessions,
don't worry too much about this.