ttools: provide stats information from BlockUploader
The BlockUploader.runMatch method now returns information about
the number of blocks truncated to the caller, rather than attempting
to inform the user about this via the logging system.
This is somewhat cleaner and allows UIs in which log messages are not
apparent to the user to convey this information in an appropriate way.

The affected stilts tasks cdsskymatch and tapskymatch now write
their own log messages summarising truncations rather than telling
BlockUploader how to do it; the behaviour of those tasks should
be unchanged by this commit.