topcat: do table sorting asynchronously
The Sort action (TopcatModel.sortBy) is now done off the EDT,
since it may be time-consuming.  While sorting is in progress,
the sort selection control is replaced with a "Sorting..." message.

Previously, a large sort on at least some OS seemed to lock up the
whole display, not just the java GUI.

The sorting has to get taken out of the CodecTest, since an EDT
may not be available at test time.