auth: fix challenge regex to cope with zero params
The regular expression used to parse www-authenticate challenges
required at least one parameter (or token68), it didn't recognise
auth-scheme-name-only challenges.  Fix it so that it does.

There are probably no legal challenges that contain only a name
and are useful to the AUTH framework, but it's more correct with
the fix, and it also enables certain questionable hacks made
during development.