votable: provide pluggable string array element sizing
Behaviour when writing VOTables with String[]-valued columns
in which the string length (ValueInfo.getElementSize()) is not known
has been improved.  The difficulty here is that the corresponding FIELD
element has to provide a fixed value for the string length,
because it's the fastest-varying dimension of a multi-dimensional
character array; it's not possible to have a column of (fixed- or
variable-length) arrays of variable-length strings (char arrays).

Up till now, in this case the VOTable writer mostly just refused to
write the column in question.  Now, behaviour is controlled by a
StringElementSizer object provided by the VOSerializerConfig object.
Default behaviour is generally to determine the string length where
required by reading all the column values in an extra pass of the
table, which inhibits streaming but is probably better than
refusing to write the column (especially e.g. in VOParquet output).
But you can configure it to do other things.