table: add RowEvaluatorTableBuilder decoder management
RowEvaluatorTableBuilder, and the table input handlers that subclass it
(CSV and ASCII), now have configuration options to select the list of
RowEvaluator.Decoder instances that can be used to identify and
decode column values.  That means that you can programmatically
or at run time choose which data types can appear in the parsed
table output.  This can be accessed for users by supplying the
"notypes" input handler config keyword with values giving the
types which are not wanted, e.g. notypes=short;float.

That can be useful if the CSV parsing is giving you shorts when you
want ints etc.