ttools: accommodate hybrid ticker edge case
Account for the case in a hybrid ticker rule where the linear threshold
is exactly at the lower or upper limit of the tick range.
Previously this resulted in divide by zeros, NaNs and AssertionErrors.
Fix it by using a dummy rule for the linear part which is not used.